refactor: mdx utils
