+++ src/app/about/grant-and-publication-materials/page.tsx
@@ -1,11 +1,11 @@
 import React from "react"
 import { Hero } from "@/components/Hero"
 import GrantAndPublicationMaterialsContent from "@/content/routes/about/grant-and-publication-materials.mdx"
-import { getMDXMetadata } from "@/lib/mdx-utils"
+import { getMDXMetadata } from "@/utils/mdx"
 
 export default async function PublicationAndGrantMaterials() {
   const metadata = getMDXMetadata(
-    "content/routes/about/grant-and-publication-materials.mdx"
+    "src/content/routes/about/grant-and-publication-materials.mdx"
   )
 
   return (
