+++ src/app/about/us/page.tsx
@@ -1,10 +1,10 @@
 import React from "react"
 import { Hero } from "@/components/Hero"
 import AboutUsContent from "@/content/routes/about/us.mdx"
-import { getMDXMetadata } from "@/lib/mdx-utils"
+import { getMDXMetadata } from "@/utils/mdx"
 
 export default async function AboutUs() {
-  const metadata = getMDXMetadata("content/routes/about/us.mdx")
+  const metadata = getMDXMetadata("src/content/routes/about/us.mdx")
 
   return (
     <>
