+++ src/components/PeopleSection.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { PeopleCard } from "@/components/card/PeopleCard"
 import { PeopleTypes, PageContentData } from "@/lib/about-types"
-import { readContentFile } from "@/lib/content-utils"
+import { readContentFile } from "@/utils/content"
 import path from "path"
 import fs from "fs/promises"
 import { cn } from "@/utils/helper"
