+++ src/content/routes/about/us.mdx
@@ -34,7 +34,7 @@ dataRefs:
 </ContentSection>
 
 <ContentSection title="Our Team">
-  <PeopleSection peopleContentPath="content/data/people.yaml" />
+  <PeopleSection peopleContentPath="src/content/data/people.yaml" />
 </ContentSection>
 
 <LocationSection title="Our Home" />
